pyproject.toml
[project]
name = "WebXiangpianbu"
version = "0.1.0"
description = "A photo album organizer"
authors = [{ name = "WAP" }]
requires-python = ">=3.11,<4"
dependencies = [
    "django==5.2.12",
    "django-pipeline==4.1.0",
    "libsasscompiler==0.1.8",
    "pillow==12.1.1",
    "pyyaml==6.0.3",
    "uwsgi==2.0.31",
    "pymemcache>=4.0.0,<5",
]

[dependency-groups]
dev = [
    "django-stubs-ext==5.2.8",
    "django-stubs[compatible-mypy]==5.2.8",
    "mypy==1.19.0",
    "playwright",
    "pytest==9.0.2",
    "pytest-django==4.11.1",
    "ruff>=0.14.8",
    "ty>=0.0.24",
    "types-pyyaml==6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["webxiangpianbu"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
plugins = ["mypy_django_plugin.main"]
exclude = '(?x)^(run/vrt-uv-cache/|run/vrt-venv/|run/vrt-home/)'

[tool.django-stubs]
django_settings_module = "webxiang.settings_ci"

[tool.ruff.format]
quote-style = "single"
